requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
fastapi==0.86.0
fastapi-cache2==0.2.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pendulum==2.1.2
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytzdata==2020.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
sqlalchemy-pagination==0.0.2
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
